requirements.txt
args==0.1.0
arrow==0.13.1
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
beautifulsoup4==4.7.1
cairocffi==1.0.2
CairoSVG==2.3.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
clint==0.5.1
constantly==15.1.0
cryptography==2.5
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
django-admin-honeypot==1.1.0
django-allauth==0.38.0
django-appconf==1.0.2
django-channels==0.7.0
django-ckeditor==5.6.1
django-compressor==2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-encrypted-model-fields==0.5.8
django-filter==2.1.0
django-haystack==2.8.1
django-hosts==3.0
django-js-asset==1.2.1
django-markdown==0.8.4
django-phonenumber-field==2.2.0
django-redis==4.10.0
django-request==1.5.4
django-softdelete==0.9.0
django-tables2==2.0.4
django-taggit==0.23.0
django-urlcrypt==0.1.6
django-weasyprint==0.5.4
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
easy-thumbnails==2.6
elasticsearch==6.3.1
googledrivedownloader==0.3
holidays==0.9.9
html5lib==1.0.1
hurry.filesize==0.9
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.12
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.1
phonenumbers==8.10.5
phonenumberslite==8.8.8
Pillow==5.2.0
pipenv==2018.11.26
pycparser==2.19
PyHamcrest==1.9.0
Pyphen==0.9.5
PySocks==1.6.8
pysolr==3.8.1
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytube==9.4.0
pytz==2018.9
rcssmin==1.0.6
redis==3.1.0
reportlab==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
serious-django-permissions==0.6
six==1.12.0
soupsieve==1.7.3
South==1.0.2
sqlparse==0.2.4
tinycss2==1.0.1
tweepy==3.7.0
txaio==18.8.1
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.5.1
WeasyPrint==45
webencodings==0.5.1
wget==3.2
Whoosh==2.7.4
zope.interface==4.6.0
